-Implemented support for using Where-Object with PrtgAPI Progress -PrtgTableCmdlets can now specify multiple Names and Ids -Implemented PrtgAPI progress for Move-Object, Set-ChannelProperty, Set-ObjectPosition, Sort-PrtgObject, Add-NotificationTrigger, Set-NotificationTrigger

+++ PrtgAPI.Tests.UnitTests/InfrastructureTests/Support/Progress/ProgressQueue.cs
@@ -67,7 +67,7 @@ namespace PrtgAPI.Tests.UnitTests.InfrastructureTests.Support.Progress
                     {
                         var maxChars = 40;
 
-                        var percentChars = ((double)snapshot[i].PercentComplete/100)*maxChars;
+                        var percentChars = (int)((double)snapshot[i].PercentComplete/100*maxChars);
 
                         var spaceChars = maxChars - percentChars;
